Remove encryptionmanager and various fixes
